Added additional option should one not want to install acrioss all repos

+++ setup_hooks.sh
@@ -11,7 +11,7 @@ gprint (){
 }
 
 rprint (){
-    echo "${RED}${1}${NORMAL}";
+    echo -e "${RED}${1}${NORMAL}\n";
 }
 
 if ! command -v pre-commit >/dev/null; then
@@ -34,6 +34,7 @@ install() {
         pre-commit init-templatedir "${PWD}";
         gprint "Now whenever a repository is cloned or created, it will have the hooks set up already!"
     fi
+    echo
 
     gprint "Do you want to install 'commit-msg-template' ? "
     gprint "If unsure what this is, goto: http://bit.ly/2My37A1  [y/n] "
@@ -44,15 +45,24 @@ install() {
             git config --global commit.template "${COMMIT_TEMPLATE}";
         fi
     fi
+    echo
 
-    find /home -name ".git" -type d -prune -exec dirname {} \; | while read -r DIR;
-    do
-        pre-commit install -f --allow-missing-config >/dev/null 2>&1;
-        if [ -f ".pre-commit-config.yaml" ]; then
-            cp .pre-commit-config.yaml "${DIR}" || true;
-        fi
+    gprint "Do you want to install 'pre-commit' across all your repos? [y/n]"
+    read -r all_repos
+    if [[ "${all_repos}" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+        find /home -name ".git" -type d -prune -exec dirname {} \; | while read -r DIR;
+        do
+            pre-commit install -f --allow-missing-config >/dev/null 2>&1;
+            if [ -f ".pre-commit-config.yaml" ]; then
+                cp .pre-commit-config.yaml "${DIR}" || true;
+            fi
+            gprint "'pre-commit' hook installed: ${DIR}";
+        done
+    else
+        gprint "Installing 'pre-commit' on your current repo."
+        pre-commit install
         gprint "'pre-commit' hook installed: ${DIR}";
-    done
+    fi
     [ $? -eq 0 ] && gprint "####################### DONE #######################"
 }
